# public/openapi.yaml
openapi: 3.0.3
info:
  title: FeeldKit API
  version: "1"
  description: |
    Field normalization, translation, and catalog HTTP API. All routes require
    `x-api-key`. Admin routes additionally require an explicit scope on the key
    (see [`docs/SDK.md`](https://github.com/feeldkit/feeldkit/blob/main/docs/SDK.md))
    and, for curator endpoints, the calling profile must be `platform_admin`.
servers:
  - url: https://feeldkit.dev/api/v1
    description: Example production host
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
...
  schemas:
...
    SimulationProfileV1:
      type: object
      required: [schema, flow_key, cases]
      properties:
        schema: { type: string, enum: ["feeldkit.simulation_profile.v1"] }
        flow_key: { type: string }
        version: { type: string }
        organization_id: { type: string, format: uuid }
        context: { type: object, additionalProperties: true }
        cases:
          type: array
          items:
            type: object
            required: [name, source_record]
            properties:
              name: { type: string }
              source_record: { type: object, additionalProperties: true }
              expected:
                type: object
                properties:
                  matched_targets: { type: array, items: { type: string } }
                  unmapped_targets: { type: array, items: { type: string } }
                  skipped_targets: { type: array, items: { type: string } }
                  status: { type: string, enum: [ok, incomplete, not_found] }
    OrgConfigProfileV1:
      type: object
      required: [schema, manifest, promotion_settings, policy_overrides, field_locks, flow_pack_overrides]
      properties:
        schema: { type: string, enum: ["feeldkit.org_config_profile.v1"] }
        manifest:
          type: object
          required: [exported_at, source_organization_id, schema_version]
          properties:
            exported_at: { type: string, format: date-time }
            source_organization_id: { type: string, format: uuid }
            feeldkit_app_version: { type: string, nullable: true }
            schema_version: { type: integer, enum: [1] }
        promotion_settings:
          type: object
          properties:
            default_scope: { type: string, enum: [org, global] }
            opt_out_global_propose: { type: boolean }
            notes: { type: string, nullable: true }
        policy_overrides:
          type: array
          items:
            type: object
            properties:
              domain: { type: string }
              matched: { type: number }
              suggested: { type: number }
              notes: { type: string, nullable: true }
        field_locks:
          type: array
          items:
            type: object
            properties:
              field_key: { type: string }
              mode: { type: string, enum: [require_review, disable_auto_apply] }
              reason: { type: string, nullable: true }
        flow_pack_overrides:
          type: array
          items:
            type: object
            properties:
              flow_key: { type: string }
              flow_pack_version: { type: string, nullable: true }
              ordinal: { type: integer, nullable: true }
              action: { type: string, enum: [skip, replace, lock, pin_version] }
              payload: { type: object, additionalProperties: true }
              notes: { type: string, nullable: true }
...
security:
  - ApiKeyAuth: []
paths:
...
  /flow/simulate:
    post:
      summary: Phase 6 — dry-run flow simulation (scope normalize + read:flows)
      description: |
        Same engine as `/flow/translate` but never persists `mapping_reviews`
        rows. Useful for pre-deployment validation against a `simulation_profile.v1`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SimulationProfileV1' }
      responses:
        "200": { description: Per-case simulation report }
        "400": { description: Invalid simulation profile }
        "404": { description: Flow not found }
...
  /admin/profile/import:
    post:
      summary: Phase 6 — apply an org_config_profile.v1 bundle to the calling org (scope admin:policies + admin:flows + admin:promotions)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profile]
              properties:
                profile: { $ref: '#/components/schemas/OrgConfigProfileV1' }
                dry_run: { type: boolean, default: false }
      responses: { "200": { description: Apply result } }